+++ IZT6ZK/Assists/ValidateInputs.cs
@@ -45,5 +45,4 @@ internal class ValidateInputs
         }
         return input;
     }
-
 }
