+++ IZT6ZK/Db/DbManager.cs
@@ -29,8 +29,6 @@ internal class DbManager : IDbManager
         Db.SaveChanges();
     }
 
-   
-
     public void DeleteQuestion(QuestionEntity question)
     {
         using var Db = new MyDbContext();
@@ -45,7 +43,6 @@ internal class DbManager : IDbManager
 
         Db.Topics.Remove(topicEntity);
         Db.SaveChanges();
-        //összekapcsolt kulcs esetén még nem lehet törölni
     }
 
     public List<QuestionEntity> SelectAllQuestions()
