+++ IZT6ZK/Models/TopicEntity.cs
@@ -9,6 +9,5 @@ internal class TopicEntity
 {
     public int TopicId { get; set; }
     public string TopicName { get; set; }
-
     public List<QuestionEntity> Questions { get; set; }
 }
